+++ example/lib/listview_page.dart
@@ -58,12 +58,10 @@ class _ListPageState extends State<ListPage> with TickerProviderStateMixin {
                       ),
                     ),
                     trailing: selectedIndex == items.indexOf(item)
-                        ? AnimatedTo(
+                        ? AnimatedTo.spring(
                             controller: _controller,
                             globalKey: GlobalObjectKey('selected_item'),
                             slidingFrom: const Offset(0, 200),
-                            duration: const Duration(milliseconds: 300),
-                            curve: Curves.easeInOut,
                             child: Container(
                               width: 12, // Smaller indicator
                               height: 12,